+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add the result of `get_config` of each module to `nodeStatusList` GraphQL API.
 - Use `set_config` of `AgentManager`, instead of `send_and_recv` in `applyNode`
   GraphQL API.
+- Use `halt` of `AgentManager`, instead of `send_and_recv` in `nodeShutdown`
+  GraphQL API.
 
 ### Removed